Type IDs are uint16 not uint8

+++ protocol/payloads/payloads.go
@@ -66,7 +66,7 @@ const (
 )
 
 type Payload interface {
-	Id() uint8
+	Id() uint16
 }
 
 func ForId(id uint16) Payload {
