+++ protocol/payloads/pl_device_set_power.go
@@ -4,4 +4,4 @@ type deviceSetPower struct {
 	Level uint16 // 0 = off; 1+ = on
 }
 
-func (deviceSetPower) Id() uint8 { return 21 }
+func (deviceSetPower) Id() uint16 { return 21 }
